+++ packages/frontend/src/core/project/LayoutHelper.test.ts
@@ -22,6 +22,6 @@ import { LayoutHelper } from './LayoutHelper';
 describe('LayoutHelper', () => {
   it('formatToPixel()', () => {
     const format = LayoutFormats.A4_PORTRAIT;
-    expect(LayoutHelper.formatToPixel(format)).toEqual({ width: 1654, height: 2339 });
+    expect(LayoutHelper.formatToPixel(format)).toEqual({ width: 2067, height: 2923 });
   });
 });